requirements.txt
supabase
google
python-dotenv
numpy
tqdm
pypdf
flask